LPD-82692 Move content-page integration tests into dedicated folder
