+++ modules/test/playwright/tests/object-web/content-page-integration/config.ts
@@ -1,0 +1,12 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+export const config = {
+	name: 'object-web.content-page-integration',
+	testDir: 'tests/object-web/content-page-integration',
+	use: {
+		testIdAttribute: 'data-qa-id',
+	},
+};
